configure passport
