+++ .env.example
@@ -1,0 +1,6 @@
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+CALLBACK_URL=
+PORT=3000
+DATABASE=
+SESSION_SECRET=
