+++ src/modules/auth/repository.js
@@ -3,12 +3,11 @@ const authCollection = require("./schema")
 
 //get user from db
 exports.getUser = async (email) => {
-
     return authCollection.findOne({ email })
 }
 
 //create and save new user to db
-exports.createNewUser = async(googleId, email, firstname, lastname, picture) => {
+exports.createNewUser = async (googleId, email, firstname, lastname, picture) => {
     try {
         let user = {
             googleId: googleId,
@@ -26,3 +25,10 @@ exports.createNewUser = async(googleId, email, firstname, lastname, picture) => 
         console.log(err)
     }
 }
+
+exports.checkIfUserExist = async (email) => {
+    let currentUser = authCollection.findOne({ email })
+    let result = currentUser ? currentUser : false
+    return result
+  }
+
