--- src/modules/event/router.js
@@ -1,8 +1,0 @@
-const eventRouter = require("express").Router()
-const { createEvent } = require("./controller")
-
-
-eventRouter.post("/event", createEvent)
-
-
-module.exports = eventRouter
